+++ index.html
@@ -7,7 +7,7 @@
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Pixelify+Sans:wght@400..700&display=swap" rel="stylesheet">
-    <script type="module" src="/script.js" defer></script>
+    <script type="module" src="js/script.js" defer></script>
     <link rel="stylesheet" href="css/animation.css">
     <link rel="stylesheet" href="css/dialogue.css">
     <link rel="stylesheet" href="css/projects.css">
