+++ js/classes/managers/cameraManagerClass.js
