+++ js/classes/managers/collisionManagerClass.js
