+++ js/classes/managers/inputManagerClass.js
