+++ js/classes/visuals/dialogueVisualsManagerClass.js
@@ -1,6 +1,6 @@
-import { ShapeDrawer } from "./shapeDrawerClass";
-import { DialogueManager } from "./dialogueManagerClass";
-import { Dialogue } from "../text/dialogue";
+import { ShapeDrawer } from "../utils/shapeDrawerClass";
+import { DialogueManager } from "../managers/dialogueManagerClass";
+import { Dialogue } from "../../../text/dialogue";
 
 export { DialogueVisualsManager }
 
@@ -20,7 +20,7 @@ const showOptions = "options-show";
 const hideOptions = "options-hide";
 
 class DialogueVisualsManager {
-    
+
     // Components
     #dialogueManager;
 
@@ -40,7 +40,7 @@ class DialogueVisualsManager {
     #inAnimation = false;
     #dialogueText;
     #textRatio = 0;
-    
+
     // Constants
     #textSpeed = 0.75;
     #hoverBaselineY;
@@ -67,7 +67,7 @@ class DialogueVisualsManager {
     }
 
     isFinishedDisplaying() {
-        return this.#textRatio >= 1 
+        return this.#textRatio >= 1
         || !this.#dialogueText // On first launch, dialogue text is not set, so its "finished displaying"
         || this.#textRatio === 0; // The start of dialogue, not done displaying
     }
@@ -97,7 +97,7 @@ class DialogueVisualsManager {
         dialogueBox.style.animationName = closeDialogue;
         clerkName.style.animationName = closeClerkName;
         nextPrompt.style.display = "none";
-        
+
         // Hide dialogue box a bit earlier than animation time to prevent one frame showing of box
         setTimeout(() => {
             dialogueBox.style.display = "none";
@@ -150,7 +150,7 @@ class DialogueVisualsManager {
 
         // Reset theta if full cycle
         if (this.#theta > Math.PI * 2) this.#theta = 0;
-        else this.#theta += delta;        
+        else this.#theta += delta;
     }
 
     displayPrompt(player, interactBox) {
@@ -176,7 +176,7 @@ class DialogueVisualsManager {
         // Create interact prompt box for far away
         this.#interactPrompt = ShapeDrawer.createInteractPromptMesh();
         this.#interactPrompt.position.set(
-            positionVector.x - ShapeDrawer.interactPromptWidth / 2, 
+            positionVector.x - ShapeDrawer.interactPromptWidth / 2,
             positionVector.y + 0.85,
             positionVector.z - ShapeDrawer.interactPromptWidth / 10
         );
@@ -211,14 +211,14 @@ class DialogueVisualsManager {
         this.#optionsOpened = false;
         this.#optionsQueued = false;
         dialogueOptions.style.animationName = hideOptions;
-        setTimeout(() => dialogueOptions.style.display = "none", 
+        setTimeout(() => dialogueOptions.style.display = "none",
             (this.#animationTimeInSeconds - 0.1) * 1000);
         setTimeout(() => this.#inAnimation = false, this.#animationTimeInSeconds * 1000);
     }
 
     #notifyPress(option) {
         if (this.#inAnimation) return; // Stop button from being pressed when its still coming in
-        if (option === "projects") 
+        if (option === "projects")
             this.#dialogueManager.switchTree(DialogueManager.treeOption.PROJECTS);
         else if (option === "interests")
             this.#dialogueManager.switchTree(DialogueManager.treeOption.INTERESTS);
